[Refactor] AdminMenuBar 함수 이동
